requirements.txt
boto3
awswrangler
google-auth==1.7.2
google-api-python-client